Removed test lockout from the login test into a new test. Renamed files to fix the order change.
