+++ OdinAPI/tests/test_4_login.py
@@ -22,6 +22,7 @@ class TestUserRoutes(unittest.TestCase):
           'password': self.data['password']
         }
         response = self.client.post('/auth/', data=json.dumps(loginInfo), content_type='application/json',  follow_redirects=True)
+        print(response.json)
         token = response.json['auth']['token'].split(' ')[1]
         self.assertEqual(response.status_code, 201)
         self.assertEqual(response.json['auth']['user'], self.data['username'])
@@ -45,14 +46,5 @@ class TestUserRoutes(unittest.TestCase):
         self.assertEqual(response.status_code, 400)
         self.assertEqual(response.json['Error'], 'Username or Password are incorrect.')
 
-    def test_lockout(self):
-        loginInfo = {
-          'username': self.data['username'],
-          'password': 'piljnvdijn'
-        }
-        for number in [1,2,3]:
-          response = self.client.post('/auth/', data=json.dumps(loginInfo), content_type='application/json',  follow_redirects=True)
-        self.assertEqual(response.status_code, 401)
-        self.assertEqual(response.json['Error'], 'Account is locked, contact administrator.')
